Fix query-utils

+++ src/app/js/query-utils.js
@@ -1,23 +1,42 @@
 import fs from "fs";
-import Database from 'better-sqlite3'
+import Database from "better-sqlite3";
 
-const openDatabase = (databaseName) => new Database(`${databaseName}.db`, { verbose: console.log });
+const openDatabase = (databaseName) =>
+  new Database(`${databaseName}.db`, { verbose: console.log });
 
 const dropDatabase = (databaseName) => {
-  fs.rmSync(`./${databaseName}.db`, { force: true }); // Remove existing
-}
+  fs.rmSync(`./${databaseName}.db`, { force: true });
+};
 
 const createDatabaseFromSql = (databaseName, sqlFilename, clean) => {
-  if (clean) {
-    dropDatabase(databaseName);
-  }
+  if (clean) dropDatabase(databaseName);
   const db = openDatabase(databaseName);
-  const sql = fs.readFileSync(sqlFilename, 'utf8');
-  db.exec(sql);
+  db.exec(fs.readFileSync(sqlFilename, "utf8"));
   db.close();
 };
 
-export {
-  createDatabaseFromSql,
-  openDatabase,
+const query = (databaseName, query, params = []) => {
+  let db = null;
+  try {
+    db = openDatabase(databaseName);
+    return db.prepare(query).get(...params);
+  } finally {
+    if (db) {
+      db.close();
+    }
+  }
 };
+
+const queryAll = (databaseName, query, params = []) => {
+  let db = null;
+  try {
+    db = openDatabase(databaseName);
+    return db.prepare(query).all(...params);
+  } finally {
+    if (db) {
+      db.close();
+    }
+  }
+};
+
+export { createDatabaseFromSql, openDatabase, query, queryAll };
